implement config handling and test command
Implements the basic module structure used for vagrant plugins and implements a basic configuration
handling. Doesn't validate the input yet and is lacking support for git sources etc. Also implement
a `vagrant puppet-library` command that starts the puppet library server with the configuration for
a given vm, so that it's possible to check the config and available modules.